+++ src/ArtifactsMMO.NET/Endpoints/Characters/Characters.cs
@@ -1,7 +1,6 @@
 ﻿using ArtifactsMMO.NET.Enums.ErrorCodes.Character;
-using ArtifactsMMO.NET.Objects;
+using ArtifactsMMO.NET.Internal;
 using ArtifactsMMO.NET.Objects.Characters;
-using ArtifactsMMO.NET.Queries;
 using ArtifactsMMO.NET.Requests;
 using ArtifactsMMO.NET.Validators;
 using System;
@@ -15,6 +14,12 @@ namespace ArtifactsMMO.NET.Endpoints.Characters
     {
         private readonly IValidator<string> _nameValidator = new CharacterNameValidator();
         public Characters(HttpClient httpClient, string apiKey) : base(httpClient, apiKey)
+        {
+        }
+
+        internal Characters(HttpClient httpClient, string apiKey,
+            IJsonSerializerOptionsFactory jsonSerializerOptionsFactory)
+            : base(httpClient, apiKey, jsonSerializerOptionsFactory)
         {
         }
 
@@ -42,11 +47,6 @@ namespace ArtifactsMMO.NET.Endpoints.Characters
         {
             _nameValidator.Validate(name);
             return await GetAsync<Character, GetCharacterError>($"characters/{name}", cancellationToken).ConfigureAwait(false);
-        }
-
-        public async Task<PagedResponse<Character>> GetAsync(CharactersQuery charactersQuery, CancellationToken cancellationToken = default)
-        {
-            return await GetAsync<Character>("characters", charactersQuery, cancellationToken).ConfigureAwait(false);
         }
     }
 }