+++ src/ArtifactsMMO.NET/Internal/JsonSerializerOptionsMode.cs
@@ -1,0 +1,8 @@
+﻿namespace ArtifactsMMO.NET.Internal
+{
+    internal enum JsonSerializerOptionsMode
+    {
+        Default,
+        Test
+    }
+}