+++ src/ArtifactsMMO.NET/Objects/Leaderboard/AccountLeaderboard.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using ArtifactsMMO.NET.Enums;
+using System.Text.Json.Serialization;
 
 namespace ArtifactsMMO.NET.Objects.Leaderboard
 {
@@ -10,16 +11,29 @@ namespace ArtifactsMMO.NET.Objects.Leaderboard
        internal AccountLeaderboard() { }
 
        [JsonConstructor]
-        internal AccountLeaderboard(string account, int achievementsPoints)
+        internal AccountLeaderboard(int position, string account, int achievementsPoints,
+           MemberStatus status)
         {
+            Position = position;
             Account = account;
             AchievementsPoints = achievementsPoints;
+            Status = status;
         }
+
+        /// <summary>
+        /// Position in the leaderboard.
+        /// </summary>
+        public int Position { get; }
 
         /// <summary>
         /// Account name.
         /// </summary>
         public string Account { get; }
+
+        /// <summary>
+        /// Member status.
+        /// </summary>
+        public MemberStatus Status { get; }
 
         /// <summary>
         /// Achievements points.