+++ tests/ArtifactsMMO.NET.Integration.Tests/Endpoints/MyCharacters/CharacterTestFixture.cs
@@ -1,0 +1,9 @@
+﻿namespace ArtifactsMMO.NET.Integration.Tests.Endpoints.MyCharacters
+{
+    public class CharacterTestFixture : TestFixture
+    {
+        private readonly string _characterName = "tst-" + new Random().Next(short.MaxValue).ToString();
+
+        public string CharacterName { get { return _characterName; } }
+    }
+}