+++ tests/ArtifactsMMO.NET.Integration.Tests/Endpoints/Resources/ResourcesEndpointTests.cs
@@ -1,4 +1,5 @@
-﻿using ArtifactsMMO.NET.Objects;
+﻿using ArtifactsMMO.NET.Integration.Tests.Helpers;
+using ArtifactsMMO.NET.Objects;
 using ArtifactsMMO.NET.Objects.Resources;
 using ArtifactsMMO.NET.Queries;
 using Microsoft.Extensions.Configuration;
@@ -21,17 +22,22 @@ namespace ArtifactsMMO.NET.Integration.Tests.Endpoints.Resources
         [Fact]
         public async Task DefaultResourcesQuery_ShouldReturnExpectedData()
         {
-            //Act
             PagedResponse<Resource>? result = await _client.Resources.GetAsync(new ResourcesQuery());
-
-
-            //Assert
 
             Assert.NotNull(result);
             Assert.True(result.Total > 0);
             Assert.NotNull(result.Data);
             var firstResource = result.Data.FirstOrDefault();
             Assert.NotNull(firstResource);
+        }
+
+        [Fact]
+        public async Task AllData_ShouldReturnExpectedData()
+        {
+            await PagedDataLoader.GetAllDataAsync(async (page, pageSize) =>
+            {
+                return await _client.Resources.GetAsync(new ResourcesQuery(page: page, size: pageSize));
+            });
         }
     }
 }